requirements.txt
python-uinput==0.10.2
wheel==0.24.0
